requirements.txt
fastapi
uvicorn
google-generativeai
python-multipart
# itsdangerous
# secure
# starlette


